LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/block-size-expected.txt


PASS Can set 'block-size' to CSS-wide keywords
FAIL Can set 'block-size' to var() references assert_equals: expected 2 but got 1
PASS Can set 'block-size' to the 'auto' keyword
FAIL Can set 'block-size' to a percent Invalid values
FAIL Can set 'block-size' to a length Invalid values
PASS Setting 'block-size' to a time throws TypeError
PASS Setting 'block-size' to an angle throws TypeError
PASS Setting 'block-size' to a flexible length throws TypeError
FAIL Setting 'block-size' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'block-size' to a URL throws TypeError
PASS Setting 'block-size' to a transform throws TypeError
PASS Can set 'min-block-size' to CSS-wide keywords
FAIL Can set 'min-block-size' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'min-block-size' to a percent Invalid values
FAIL Can set 'min-block-size' to a length Invalid values
PASS Setting 'min-block-size' to a time throws TypeError
PASS Setting 'min-block-size' to an angle throws TypeError
PASS Setting 'min-block-size' to a flexible length throws TypeError
FAIL Setting 'min-block-size' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'min-block-size' to a URL throws TypeError
PASS Setting 'min-block-size' to a transform throws TypeError
PASS Can set 'max-block-size' to CSS-wide keywords
FAIL Can set 'max-block-size' to var() references assert_equals: expected 2 but got 1
PASS Can set 'max-block-size' to the 'none' keyword
FAIL Can set 'max-block-size' to a percent Invalid values
FAIL Can set 'max-block-size' to a length Invalid values
PASS Setting 'max-block-size' to a time throws TypeError
PASS Setting 'max-block-size' to an angle throws TypeError
PASS Setting 'max-block-size' to a flexible length throws TypeError
FAIL Setting 'max-block-size' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'max-block-size' to a URL throws TypeError
PASS Setting 'max-block-size' to a transform throws TypeError
